python/pyproject.toml
[tool.poetry]
name = "grpc-vimn-python"
version = "0.1.0"
description = ""
authors = ["Miron Olszewski <miron.olszewski@vimn.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.39.0"
grpcio-tools = "^1.39.0"
click = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
